@@ -3,13 +3,13 @@ apply plugin: 'com.github.dcendents.android-maven'//this
 group='com.github.ZhouHoubin'//this
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
 
     defaultConfig {
         minSdkVersion 19
-        targetSdkVersion 28
-        versionCode 45
-        versionName "4.5"
+        targetSdkVersion 29
+        versionCode 46
+        versionName "4.6"
         //21:修复文件追加第一行直接覆盖的情况
         //29:增加UIDump
         //30:增加悬浮窗
@@ -27,8 +27,8 @@ android {
 
 dependencies {
     compileOnly  'de.robv.android.xposed:api:82'
-    implementation 'com.android.support:support-annotations:28.0.0'
-    implementation 'com.android.support:appcompat-v7:28.0.0'
+    implementation 'androidx.annotation:annotation:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
     implementation files('libs/lite-orm-1.9.2.jar')
     implementation 'com.github.czy1121:settingscompat:1.1.4'
 }
